requirements.txt
langgraph==0.2.56
langchain==0.3.7
langchain-openai==0.2.8
langchain-anthropic==0.3.0
langchain-community==0.3.7
streamlit==1.40.2
chromadb==0.5.23
onnxruntime>=1.17.0
pandas==2.2.3
pypdf==5.1.0
python-dotenv==1.0.1
pydantic==2.10.3
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.28.1
